rename framework -> integration (more consistent with docs) (#128)
