+++ tox.ini
@@ -132,7 +132,7 @@ exclude =
     wandb/wandb_torch.py,
     wandb/data/,
     wandb/errors/,
-    wandb/framework/,
+    wandb/integration/,
     wandb/old/,
     wandb/sdk_py27/,
     wandb/proto/,
