+++ wandb/fastai/__init__.py
@@ -2,9 +2,9 @@
 Compatibility fastai module.
 
 In the future use:
-    from wandb.framework.fastai import WandbCallback
+    from wandb.integration.fastai import WandbCallback
 """
 
-from wandb.framework.fastai import WandbCallback
+from wandb.integration.fastai import WandbCallback
 
 __all__ = ['WandbCallback']
