--- wandb/framework/keras/__init__.py
@@ -1,7 +1,0 @@
-"""
-wandb framwwork keras module.
-"""
-
-from wandb.framework.keras.keras import WandbCallback
-
-__all__ = ['WandbCallback']
