+++ wandb/integration/keras/__init__.py
@@ -1,0 +1,7 @@
+"""
+wandb integration keras module.
+"""
+
+from .keras import WandbCallback
+
+__all__ = ['WandbCallback']
