+++ wandb/keras/__init__.py
@@ -2,9 +2,9 @@
 Compatibility keras module.
 
 In the future use:
-    from wandb.framework.keras import WandbCallback
+    from wandb.integration.keras import WandbCallback
 """
 
-from wandb.framework.keras import WandbCallback  # type: ignore
+from wandb.integration.keras import WandbCallback  # type: ignore
 
 __all__ = ['WandbCallback']
