+++ wandb/lightgbm/__init__.py
@@ -2,9 +2,9 @@
 Compatibility lightgbm module.
 
 In the future use:
-    from wandb.framework.lightgbm import wandb_callback
+    from wandb.integration.lightgbm import wandb_callback
 """
 
-from wandb.framework.lightgbm import wandb_callback
+from wandb.integration.lightgbm import wandb_callback
 
 __all__ = ['wandb_callback']
