+++ wandb/xgboost/__init__.py
@@ -2,9 +2,9 @@
 Compatibility xgboost module.
 
 In the future use:
-    from wandb.framework.xgboost import wandb_callback
+    from wandb.integration.xgboost import wandb_callback
 """
 
-from wandb.framework.xgboost import wandb_callback
+from wandb.integration.xgboost import wandb_callback
 
 __all__ = ['wandb_callback']
